_featsSingleFrame returns a vector of bin indices instead of raw pow of magnitudes.

+++ src/spectral_analyzer.hpp
@@ -14,8 +14,8 @@ private:
     std::vector<std::vector<float>> _split(const std::vector<float>& signal) const;
     void                            _hannSingleFrame(std::vector<float>& frame) const;
     void                            _hann(std::vector<std::vector<float>>& split_signal) const;
-    std::vector<float>              _featsSingleFrame(std::vector<float>& frame);
-    std::vector<std::vector<float>> _feats(std::vector<std::vector<float>>& split_signal);
+    std::vector<int>                _featsSingleFrame(std::vector<float>& frame);
+    std::vector<std::vector<int>>   _feats(std::vector<std::vector<float>>& split_signal);
 
 public:
     SpectralAnalyzer(const float frame_duration, const int sampling_frequency, const int feature_ratio);
